Addressing Bhroam's comment

+++ src/include/libutil.h
@@ -298,7 +298,7 @@ void convert_duration_to_str(time_t duration, char* buf, int bufsize);
  * deduce the preemption ordering to be used for a job
  */
 struct preempt_ordering *
-get_preemption_order(struct preempt_ordering *porder, double req, double used);
+get_preemption_order(struct preempt_ordering *porder, int req, int used);
 
 #ifdef  __cplusplus
 }
